Enforce serializable cursor
TL;DR: This enforces that cursors can only be composed of Objects of
built-in JSON compatible classes: Array, FalseClass, Float, Hash,
Integer, NilClass, String, and TrueClass.

Both Resque and Sidekiq serialize job arguments as JSON. Because
`cursor_position` is not part of the ActiveJob arguments, it is subject
to this serialization method.

Objects other than the basic types JSON supports are serialized by
calling `.to_s`, which means they are subject to one way serialization;
once serialized, there is no way (out of the box) for them to be
deserialized.

One such example is Time objects, which are serialized to a reasonable
String representation, but when deserialized simply result in a String
containing that textual representation.

Sneakily enough, because ActiveSupport "augments" `Time#==` adding
automatic coercion, `time == JSON.parse(JSON.dump(time))`, which means
it isn't always noticed in tests.

Nonetheless, this can result in unexpected errors, especially since it
only "matters" if the job is interrupted (as otherwise the cursor is
never serialized.

Therefore, we should enforce that the (de)serialization round trip
results in the same cursor it was given without corruption. The simplest
way to do this is to traverse the cursor looking for unsupported
classes. Since cursors should typically be very small, this should not
add any significant performance penalty.