Fix enumerator #size for explicit column select
Use `count(:all)` instead of [`count`][2] to get the number of items in
a `ActiveRecordEnumerator` and in a `ActiveRecordBatchEnumerator`. This
prevents SQL syntax errors when the relation has an explicit column
selection with `select(:col1, :col2)`, which would attempt to run the
invalid:

```sql
SELECT COUNT(col1, col2) FROM table;
```

Note that this is very similar to what [ActiveRecord::Relation#size][1]
does, without the "smart" logic to return the length of the array of
records already loaded. This keeps the logic in this gem simpler, and
closer to what we had before with regular `count()`.

The issue surfaced [in maintenance_tasks][3], where `#size` is called
for a `ActiveRecordBatchEnumerator` with tasks not defining a custom
`count` method. This would fail for batches selecting a subset of
columns, such as `Product.select(:id, :name).in_batches`

[1]:https://api.rubyonrails.org/classes/ActiveRecord/Relation.html#method-i-size
[2]:https://api.rubyonrails.org/classes/ActiveRecord/Calculations.html#method-i-count
[3]:https://github.com/Shopify/maintenance_tasks/blob/55a5fd3765dd312ba85eaad47b6e8e81feb566a5/app/jobs/concerns/maintenance_tasks/task_job_concern.rb#L107