Easier intro for custom enumerator docs
When using this doc to explain how this gem works I noticed it's uninituitive at first that Iteration uses the persisted cursor position to find objects that come _after_ the object last processed before exiting. It's much easier to grok by showing the array enumerator first, then the more complex Stripe example, and leave the Redis example as a cool trick at the end (instead of leading with it).