Extract linting into separate CI job
Because we can simply set `TargetRubyVersion` to the version whose
syntax we want to support, we can run the Rubocop in a single job,
rather than repeating it across all jobs (and matrix combinations).

A side benefit is that this separates test failures from linting
failures, whereas previously tests would be skipped if the linter
failed.