Refactor gemfiles
We have several gemfiles in which we have specific version requirements
for `activejob` and `activerecord`, for use in our CI matrix.

The current approach breaks on Ruby HEAD because it is detected as
specifying multiple versions/sources for those gems. Instead, this
refactors them to use an instance variable approach, similar to that
used in Shopify/maintenance_tasks.