Define methods normally in `IntegrationBehavior`
This changes the `private` helper methods from being defined in the
`included` block to being defined as regular methods in the module.

`included` should only be used for using class macros.

This fixes method redefinition warnings, because due to the `included`
approach, we were defining the methods on the host class (instead of
this module), which would go on to redefined them.